+++ src/lang/mk.js
@@ -80,6 +80,7 @@ module.exports = {
   not_in: "Избраното поле :attribute е невалидно.",
   not_regex: "The :attribute format is invalid.",
   numeric: "Полето :attribute мора да биде број.",
+  object: "Полето :attribute мора да биде низа.",
   present: "The :attribute field must be present.",
   regex: "Полето :attribute е во невалиден формат.",
   required: "Полето :attribute е задолжително.",