+++ src/lang/ms.js
@@ -80,6 +80,7 @@ module.exports = {
   not_in: ":attribute tidak sah.",
   not_regex: "Format :attribute adalah tidak sah.",
   numeric: ":attribute mesti nombor.",
+  object: ":attribute mesti jujukan.",
   present: ":attribute mesti wujud.",
   regex: "Format :attribute tidak sah.",
   required: "Ruangan :attribute diperlukan.",